perf(ui): defer connection bootstrap scheduler
